Merge firmetix branch version 1.0.0 to main

+++ alimata/sensors/potentiometer.py
@@ -32,7 +32,7 @@ class Potentiometer(Sensor):
         super().__init__(board=board, pin=pin, type_=PIN_MODE.ANALOG_INPUT, differential=differential,
                          on_change=on_change)
 
-    def mapped_data(self, min_val: int, max_val: int) -> int:
+    def mapped_data(self, min_val: float, max_val: float) -> float:
         """Return the current value of the Potentiometer mapped to the given range (min to max)"""
         return map_range(self.__data, 0, 255, min_val, max_val)
